+++ src/main/java/org/jabref/gui/preferences/customization/CustomizationTab.java
@@ -1,6 +1,7 @@
 package org.jabref.gui.preferences.customization;
 
 import javafx.fxml.FXML;
+import javafx.scene.control.Button;
 import javafx.scene.control.CheckBox;
 import javafx.scene.control.TextField;
 
@@ -14,6 +15,9 @@ public class CustomizationTab extends AbstractPreferenceTabView<CustomizationTab
 
     @FXML private CheckBox useCustomDOI;
     @FXML private TextField useCustomDOIName;
+    @FXML private CheckBox useCustomSpringerApiKey;
+    @FXML private TextField useCustomSpringerApiKeyName;
+    @FXML private Button checkSpringerApiKeyButton;
 
     public CustomizationTab() {
         ViewLoader.view(this)
@@ -26,11 +30,28 @@ public class CustomizationTab extends AbstractPreferenceTabView<CustomizationTab
         return Localization.lang("Customization");
     }
 
+    /**
+     * Initialize CustomizationTab
+     */
     public void initialize() {
         this.viewModel = new CustomizationTabViewModel(dialogService, preferencesService);
 
         useCustomDOI.selectedProperty().bindBidirectional(viewModel.useCustomDOIProperty());
         useCustomDOIName.textProperty().bindBidirectional(viewModel.useCustomDOINameProperty());
         useCustomDOIName.disableProperty().bind(useCustomDOI.selectedProperty().not());
+
+        useCustomSpringerApiKey.selectedProperty().bindBidirectional(viewModel.useCustomSpringerKeyProperty());
+        useCustomSpringerApiKeyName.textProperty().bindBidirectional(viewModel.useCustomSpringerKeyNameProperty());
+        useCustomSpringerApiKeyName.disableProperty().bind(useCustomSpringerApiKey.selectedProperty().not());
+
+        checkSpringerApiKeyButton.disableProperty().bind(useCustomSpringerApiKey.selectedProperty().not());
+    }
+
+    /**
+     * Trigger check Springer API Key when clicking the check connection button
+     */
+    @FXML
+    void checkSpringerApiKey() {
+        viewModel.checkSpringerApiKey();
     }
 }