+++ src/main/java/org/jabref/preferences/PreferencesService.java
@@ -40,6 +40,7 @@ import org.jabref.logic.net.ProxyPreferences;
 import org.jabref.logic.openoffice.OpenOfficePreferences;
 import org.jabref.logic.preferences.DOIPreferences;
 import org.jabref.logic.preferences.OwnerPreferences;
+import org.jabref.logic.preferences.SpringerApiKeyPreferences;
 import org.jabref.logic.preferences.TimestampPreferences;
 import org.jabref.logic.protectedterms.ProtectedTermsPreferences;
 import org.jabref.logic.remote.RemotePreferences;
@@ -149,7 +150,11 @@ public interface PreferencesService {
 
     DOIPreferences getDOIPreferences();
 
+    SpringerApiKeyPreferences getSpringerAPIKeyPreferences();
+
     void storeDOIPreferences(DOIPreferences preferences);
+
+    void storeCustomSpringerKeyPreferences(SpringerApiKeyPreferences preferences);
 
     OwnerPreferences getOwnerPreferences();